Use the env token provider if one is not specified
